+++ src/configurator/validators.js
@@ -1,5 +1,5 @@
-export offline from "./validators/offline.js"
 export assets from "./validators/assets.js"
 export baseUrl from "./validators/baseUrl.js"
-export production from "./validators/production.js"
 export devPort from "./validators/devPort.js"
+export offline from "./validators/offline.js" // ! must be after baseUrl
+export production from "./validators/production.js"